+++ src/main/java/com/spring/fleamarket/domain/account/mapper/AccountFindMapper.java
@@ -1,5 +1,7 @@
 package com.spring.fleamarket.domain.account.mapper;
 
+import java.util.List;
+
 import com.spring.fleamarket.domain.model.Account;
 import com.spring.fleamarket.domain.model.AccountImage;
 
@@ -8,5 +10,7 @@ public interface AccountFindMapper {
 	public Account selectAccountByName(String name);
 	
 	public AccountImage selectAccountImageByAccountId(int accountId);
+	
+	public List<Account> selectAll();
 	
 }
