+++ src/main/java/com/spring/fleamarket/domain/account/service/impl/AccountFindServiceImpl.java
@@ -1,10 +1,13 @@
 package com.spring.fleamarket.domain.account.service.impl;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.spring.fleamarket.domain.account.mapper.AccountFindMapper;
 import com.spring.fleamarket.domain.account.service.AccountFindService;
+import com.spring.fleamarket.domain.model.Account;
 import com.spring.fleamarket.domain.model.AccountImage;
 
 @Service
@@ -16,6 +19,11 @@ public class AccountFindServiceImpl implements AccountFindService {
 	@Override
 	public AccountImage selectAccountImageByAccountId(int accountId) {
 		return mapper.selectAccountImageByAccountId(accountId);
+	}
+
+	@Override
+	public List<Account> selectAll() {
+		return mapper.selectAll();
 	}
 
 }
